Create graph draw generator with penalty pullups
To support WSDC, the penalties for pulling up a team need to be decided
with edge weights within the graph, rather than computed before passing
the brackets to the graph. For this commit, a new draw generator was
created and the draw strength by ranks penalty calculated. Therefore,
all the teams are put in the same graph, and the pullup flag added
after-the-fact.

The maximal side imbalance constraint was improved, and penalties are no
longer recalculated twice for each edge.