Create new email status for failure to send
This commit will attempt to add more visibility to the email sending
process, where some emails seem to fail and that makes the rest of the
batch also fail with no record of the emails created.

To make sure we only fail on specific emails, we can send each email
one-at-a-time (keeping the same connection for performance). Then, we
can catch failures and create a status object for it with the exception.